requirements.txt
torch~=1.11.0
timm~=0.5.4
torchvision~=0.12.0
numpy~=1.21.1
pandas~=1.1.5
Pillow~=9.0.1
scikit-learn~=1.0.2
tqdm~=4.62.3
scikit-image~=0.19.2
PyYAML~=6.0